Allow plugin/theme developers to filter sections/settings
I realize that there is a method you suggest in the documentation of your plugin - updating the option in the database. However, I feel that allowing for a filter (especially with the way that the options array is structured) provides for a better way to programmatically create options on the fly.

This method allows me to utilize both programmatic addition of options as well as adding options in the standard UI method.